Test were failing when we try to unmount a already unmounted device. Now, Unmount will only be done when mountpoint is there otherwise it will skip unmount and test will not fail even if somehow this script is called for already unmounted device

+++ testcases/ddt/scripts/blk/blk_device_umount.sh
@@ -55,9 +55,10 @@ done
 test_print_trc "Umounting device"
 test_print_trc "MNT_POINT: $MNT_POINT"
 
-do_cmd "umount $MNT_POINT"
-do_cmd "rm -rf $MNT_POINT"
-
+if mountpoint $MNT_POINT; then
+  do_cmd "umount $MNT_POINT"
+  do_cmd "rm -rf $MNT_POINT"
+fi
